api/requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.3
python-dotenv==1.0.0
